fix(table): Allow display of additional table row, if row height > 1
Prior to this commit, a table would only display rows whose height
could entirely fit within the available area.  As a result, table rows
with `row.height > 1`, which require multiple terminal rows to
display, could result in empty rows being rendered at the bottom of
the table's display area.

This commit updates the rendering of `Table` to also include rows that
may be partially visible within the available area.  That is, even if
the display area does not have sufficient height to fully display an
additional row of the table, it may still display the first few lines
of an additional table row.

This change does not affect the change of the `offset`, which adjusts
the viewing bounds to ensure that the selected row is displayed.  This
is deliberate, since view should be scrolled to allow the selected row
to be displayed in its entirety.