fix(buffer): avoid overflow while getting the position of indexes
This commit fixes an unsafe conversion issue in the `pos_of` method of
Buffer where the final position is silently truncated if an index larger
than u16::MAX is provided. It was due to the unchecked cast using
"as u16" and it has been replaced with a proper u16::try_from call.

`pos_of` now panics if the given position is still not inside the
available calculation range, which shouldn't be a problem in most of
terminals.

Related to #1449
Coded on livestream: https://www.youtube.com/live/jiTtSmdRNd4