fix(canvas): Draw default-colored blocks with Marker::HalfBlock
Prior to this commit, the `HalfBlockGrid` used `Color::Reset` to
identify pixels that did not have anything drawn in them.  As a
result, elements drawn with `Color::Reset` would be skipped.  This
behavior is inconsistent with other implementations of
`widgets::canvas::Grid`, which draw elements of `Color::Reset` using
the default foreground color.

This commit updates the implementation of `HalfBlockGrid` to store
pixels of `Option<Color>`, rather than `Color`.  This allows the
implementation of `Grid::save` for `HalfBlockGrid` to distinguish
between cells that have not been painted (`None`) and cells that have
been painted with the default foreground color (`Some(Color::Reset)`).