feat(chart): Allow Braille/Dot charts to overlap with Block
The `Chart` widget allows multiple line/scatter plots to be drawn on
the same axis.  Prior to this commit, for a given tile of the canvas,
the output was entirely determined by the top-most line plot to touch
that tile.

In most cases, two overlapping lines cannot be rendered on top of each
other.  For example, drawing `Marker::Braille` on top of `Marker::Dot`
requires replacing the character being drawn.  However,
`Marker::Block` is an exception, as a block may be accurately
represented as either `symbols::block::FULL` in the foreground,
or with a background color.

This commit updates the rendering of the `Chart` widget to allow two
line/scatter plots to overlap, where the first chart uses
`Marker::Block`, and the second chart uses either `Marker::Braille` or
`Marker::Dot`.