Make Size and Position generic
This change makes it possible to just use Size and position as the types
that are used to modify existing sizes / positions etc. Instead of
adding specific types for Offset and SizeDelta, Position and Size now
take generic parameters which can be signed integers. By default these
are u16, to be compatible with the existing behavior of all existing
code. Offset is now a type alias `Position<i32>`. Existing code works
without modification.