feat: cell diff options
Problem: Escape sequences always cause a cell to count as "multiwidth",
even when it doesn't render wider than one cell, or not as wide as the
escape sequence would be computed as.

Solution: Convert `skip:bool` to enum. Add enum option `ForceWidth` to
force a cell width for diffing.

When using the option, this also fixes some bug where diffing is not
idempotent and causes a diff operation for `(symbol.len() - 1)` times.

There are three test cases:

1. Rendering hyperlinks by squeezing the escape sequence into the first
   cell and forcing the width to the unicode width of the text part.
   This is much easier to implement for a Link widget, as it would only
   need to get the unicode-width once and not iterate over graphemes
   like Spans must do.
2. Rendering hyperlinks by squeezing the opening sequence into the first
   cell with the first grapheme and forcing the width to that of the
   first grapheme. Then rendering each grapheme as usual. Then squeezing
   the closing sequence into the last cell with the last grapheme and
   forcing the width to that of the last grapheme.
   This is harder to implement for a Link widget, as it would have to
   iterate over graphemes with their width like Spans do.
3. Kitty image sequence with utf-8 placeholders, similar to 2 but with
   known constant grapheme widths.