Add support for quadrant & sextant symbols
`TwoByFourGrid` (formerly `BrailleGrid`) was renamed again to
`PatternGrid` and made generic to support any pixel-like pattern.

Support was added for quadrant and sextant symbols.

The `symbols` module was reorganized with all pseudo-pixel symbols now
under a single `pixel` submodule. For consistency, the Braille module was
modified to contain a similar array of symbols and the `decoder` generic
parameter was removed to use character lookup tables directly.