feat(canvas)!: Add quadrant, sextant and octant markers (#2235)
The octant marker is an alternative to the Braille marker with the same
resolution, but offering densely packed, regular pseudo-pixels, without
visible bands between rows and columns. 

Quadrant and Sextants are also added to support 2x2 and 2x3.

Sextant and Octant unicode characters that are less widely supported at
the moment, which is why `Braille` was left as the default.

BREAKING CHANGE: addition of new variants to `Marker` and removal of
no longer used constants in `ratatui::symbols::braille`.